replace metrics crates
New metrics crate now has feature parity with the legacy crate and
can be used to replace it in actual usage.

This change removes the legacy crate and replaces it with the new
one.
