+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
     "histogram",
     "logger",
     "metrics",
-    "metrics2",
     "ratelimiter",
     "streamstats",
     "timer",
