--- metrics/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Unreleased
-
-# 1.0.0 - 2019-12-13
-
-Initial release.
