--- metrics/rustfmt.toml
@@ -1,5 +1,0 @@
-reorder_imports = true
-reorder_modules = true
-use_field_init_shorthand = true
-use_try_shorthand = true
-
