--- metrics/src/common/mod.rs
@@ -1,21 +1,0 @@
-// Copyright 2019 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0
-// http://www.apache.org/licenses/LICENSE-2.0
-
-mod measurement;
-mod output;
-mod percentile;
-mod point;
-mod reading;
-mod source;
-mod statistic;
-mod summary;
-
-pub use measurement::Measurement;
-pub use output::Output;
-pub use percentile::Percentile;
-pub use point::Point;
-pub use reading::Reading;
-pub use source::Source;
-pub use statistic::Statistic;
-pub use summary::Summary;
