--- metrics2/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rustcommon-metrics2"
-version = "0.0.0"
-authors = ["Brian Martin <bmartin@twitter.com>"]
-edition = "2018"
-license = "Apache-2.0"
-description = "Metrics library with derived summary statistics"
-homepage = "https://github.com/twitter/rustcommon/metrics2"
-repository = "https://github.com/twitter/rustcommon"
-
-[dependencies]
-dashmap = "3.11.4"
-rustcommon-atomics = { path = "../atomics" }
-rustcommon-heatmap = { path = "../heatmap" }
-rustcommon-streamstats = { path = "../streamstats" }
