--- metrics2/src/summary/mod.rs
@@ -1,9 +1,0 @@
-// Copyright 2020 Twitter, Inc.
-// Licensed under the Apache License, Version 2.0
-// http://www.apache.org/licenses/LICENSE-2.0
-
-mod summary;
-mod summary_type;
-
-pub use summary::Summary;
-pub(crate) use summary_type::SummaryType;
